reset password - cleanup, remove custom loading message screen
